requirements.txt
pandas
plotly
psutil
mpl_finance
matplotlib
munch
mysql-connector-python
python-dotenv